[explorer/rest] refactor: move Symbol block format helpers to shared format module
